rp2040: fix timer wrap deadlock in ws2812 vendor driver
The current implementation could be deadlocked if the last update was
more then ~35min ago. This is due to the time resolution for the RP2040
on ChibiOS being 32bit and the counter running at 1MHz and
timer_expired32() only handling timer intervals up to (UINT32_MAX / 2).
This can easily happen if the keyboard is put to sleep or the RGB being
disabled for this time.

The solution is to use the full 64bit time resolution of the RP2040
which wraps after ~584942 years - let me know if you happen to witness
this. Unfortunately we have fallback to the pico-sdk implementation as
there is no 64bit timer resolution support in the ChibiOS implementation
of the RP2040.